Split the Telomare god-module into IR vocabulary modules
Telomare.hs (1100 lines: every IR, every error type, the Term3Builder
DSL) becomes seven focused modules:

  Telomare.IR.Loc      source positions, LocTag, LocatedName
  Telomare.IR.Base     shared base functors, embed/extract classes,
                       constraint-polymorphic pattern synonyms
  Telomare.IR.Types    DataType / PartialType
  Telomare.IR.Surface  UnprocessedParsedTermF, UPT/AUPT, patterns
  Telomare.IR.Core     Term1/2/3, StuckExpr, CompiledExpr, RunTimeError,
                       TelomareLike/AbstractRunTime
  Telomare.IR.Builder  Term3Builder lowering DSL
  Telomare.Error       per-stage error types unioned by EvalError

Telomare remains as a re-export facade for one commit so the partition
is verified acyclic before importers are rewritten.
