+++ src/Telomare/IR/Loc.hs
@@ -1,0 +1,101 @@
+{-# LANGUAGE DeriveAnyClass #-}
+{-# LANGUAGE DeriveGeneric  #-}
+{-# LANGUAGE LambdaCase     #-}
+
+-- |Source locations. The lowest layer of the IR vocabulary: every stage
+-- annotates its terms with a 'LocTag' so diagnostics can say where a term
+-- came from.
+module Telomare.IR.Loc where
+
+import Control.DeepSeq (NFData (..))
+import Data.GenValidity (GenValid)
+import Data.Validity (Validity)
+import GHC.Generics (Generic)
+
+data SourcePosition = SourcePosition
+  { sourcePositionLine   :: Int
+  , sourcePositionColumn :: Int
+  , sourcePositionOffset :: Int
+  }
+  deriving (Eq, Show, Ord, Generic, NFData)
+
+instance Validity SourcePosition
+instance GenValid SourcePosition
+
+data SourceSpan = SourceSpan
+  { sourceSpanFile  :: Maybe FilePath
+  , sourceSpanStart :: SourcePosition
+  , sourceSpanEnd   :: SourcePosition
+  }
+  deriving (Eq, Show, Ord, Generic, NFData)
+
+instance Validity SourceSpan
+instance GenValid SourceSpan
+
+data LocTag
+  = SourceLoc SourceSpan
+  | GeneratedLoc String (Maybe LocTag)
+  | BuiltinLoc String
+  | RuntimeLoc
+  | DecompiledLoc
+  | UnknownLoc
+  deriving (Eq, Show, Ord, Generic, NFData)
+
+instance Validity LocTag
+instance GenValid LocTag
+
+locStartLineColumn :: LocTag -> Maybe (Int, Int)
+locStartLineColumn = \case
+  SourceLoc span ->
+    let start = sourceSpanStart span
+    in Just (sourcePositionLine start, sourcePositionColumn start)
+  GeneratedLoc _ parent -> parent >>= locStartLineColumn
+  _ -> Nothing
+
+renderLocTag :: LocTag -> Maybe String
+renderLocTag loc = case locStartLineColumn loc of
+  Just (line, column) -> Just $ "line " <> show line <> ", column " <> show column
+  Nothing             -> Nothing
+
+-- |`file:line:column` when the location knows which file it came from,
+-- `line:column` when it doesn't. Compact enough to list many of them.
+renderLocTagCompact :: LocTag -> Maybe String
+renderLocTagCompact = \case
+  SourceLoc spn ->
+    let start = sourceSpanStart spn
+        lineCol = show (sourcePositionLine start) <> ":" <> show (sourcePositionColumn start)
+    in Just $ maybe lineCol (<> (":" <> lineCol)) (sourceSpanFile spn)
+  GeneratedLoc _ parent -> parent >>= renderLocTagCompact
+  _ -> Nothing
+
+renderLocTagVerbose :: LocTag -> String
+renderLocTagVerbose loc = case (loc, renderLocTag loc) of
+  (GeneratedLoc reason _, Just rendered) -> rendered <> " (generated by " <> reason <> ")"
+  (GeneratedLoc reason Nothing, _)       -> "generated by " <> reason
+  (BuiltinLoc name, _)                   -> "builtin " <> show name
+  (RuntimeLoc, _)                        -> "runtime value"
+  (DecompiledLoc, _)                     -> "decompiled value"
+  (UnknownLoc, _)                        -> "unknown location"
+  (_, Just rendered)                     -> rendered
+  (_, Nothing)                           -> "unknown location"
+
+newtype LocatedName = LocatedName (LocTag, String)
+  deriving (Eq, Ord, Show)
+
+locatedNameLoc :: LocatedName -> LocTag
+locatedNameLoc (LocatedName (loc, _)) = loc
+
+locatedNameText :: LocatedName -> String
+locatedNameText (LocatedName (_, name)) = name
+
+locatedName :: LocTag -> String -> LocatedName
+locatedName loc name = LocatedName (loc, name)
+
+letBindingName :: (LocatedName, a) -> String
+letBindingName (name, _) = locatedNameText name
+
+letBindingValue :: (LocatedName, a) -> a
+letBindingValue (_, value) = value
+
+letBindingLoc :: (LocatedName, a) -> LocTag
+letBindingLoc (name, _) = locatedNameLoc name
